add the code change to handle the schema sent through requestedColumnIds
